+++ packages/edit-site/src/components/list/table.js
@@ -89,6 +89,7 @@ export default function Table( { templateType } ) {
 									params={ {
 										postId: template.id,
 										postType: template.type,
+										canvas: 'edit',
 									} }
 								>
 									{ decodeEntities(